requirements.txt
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
requests==2.32.3
schedule==1.2.2
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.1.3
